requirements.txt
telethon
python-dotenv
yt-dlp
aiohttp
